Aligning VICII. Sprites and graphics are, so far, drawn only when the vertical raster is in the "screen" zone of the window, but they should be draw within the visible zone (even if they are later hidden by the border) to be able to detect collisions weven when they are not seen. Many changes have been done in the way the different methods to draw and simule work, but they are not still working. Additionally some modifications in the datasette have been added. Now to build up a peripheral a reference to the computer is needed. Many other games (CRT format have vbeen downloaded and clasified between the ones working and the ones not still worjing).

+++ include/CORE/Raster.hpp
@@ -156,13 +156,13 @@ namespace MCHEmul
 	{
 		bool result = false;
 
-		int cP = (int)_currentPosition_0;
-		cP += i; // Can move to the next (o nexts) lines...
+		int cP = (int) _currentPosition_0;
+		cP += (int) i; // Can move to the next (o nexts) lines...
 		if (result = (cP >= (int) _maxPositions))
 			while (cP >= (int) _maxPositions)
 				cP -= (int) _maxPositions;
 
-		cP += (int) _firstPosition;
+		cP += (int) _firstPosition; // Now is no longer in base 0...
 		if (cP >= (int) _maxPositions)
 			cP -= (int) _maxPositions;
